+++ cad_curves.html
@@ -3,18 +3,11 @@
 	<head>
 		<meta charset="utf-8">
 		<title>NURBS Interpolation</title>
-		<!-- <script src="./Rendering/three.module.js"></script>
-		<script src="./Rendering/OrbitControls.js"></script> -->
 		<script src="./libs/stats.min.js"></script>
 		<script src="./libs/rhino3dm.js"></script>
 		<script type="module" src="./cad_curves.js"></script>
 		<link type="text/css" href="css/style.css" rel="stylesheet">
 	</head>
 	<body>
-		<!-- <script src="./React/react.production.min.js"></script>
-		<script src="./React/react-dom.production.min.js"></script>
-		<script src="./React/babel.min.js"></script>
-		<script type="text/babel" src="react_test.js"></script>
-		<div id="root"></div> -->
 	</body>
 </html>
